requirements.txt
# do this : pip install -r requirements.txt

streamlit
fastapi
uvicorn
requests
# openai-whisper
# torch
openai
dotenv
python-docx
python-multipart
fpdf
